+++ rtsim/src/data/npc.rs
@@ -6,7 +6,10 @@ use crate::{
 pub use common::rtsim::{NpcId, Profession};
 use common::{
     character::CharacterId,
-    comp,
+    comp::{
+        self,
+        agent::{PidGain, PidMode},
+    },
     grid::Grid,
     resources::Time,
     rtsim::{
@@ -78,6 +81,26 @@ impl Controller {
 
     pub fn do_goto(&mut self, wpos: Vec3<f32>, speed_factor: f32) {
         self.activity = Some(NpcActivity::Goto(wpos, speed_factor));
+    }
+
+    /// go to with height above terrain and direction
+    pub fn do_goto_with_height_and_dir(
+        &mut self,
+        wpos: Vec3<f32>,
+        speed_factor: f32,
+        height: Option<f32>,
+        dir: Option<Dir>,
+        pid_mode: PidMode,
+        pid_gain: PidGain,
+    ) {
+        self.activity = Some(NpcActivity::GotoFlying(
+            wpos,
+            speed_factor,
+            height,
+            dir,
+            pid_mode,
+            pid_gain,
+        ));
     }
 
     pub fn do_gather(&mut self, resources: &'static [ChunkResource]) {